fixing xml and rearranging searching stuff
