+++ js/data_input/load.js
@@ -1,4 +1,4 @@
-define(['require', 'jquery', 'data_input/liveHTML', 'data_input/array', 'data_input/HTML', 'data_input/JSON', 'data_input/xml'],function (require) {
+define(['require', 'jquery', 'data_input/liveHTML', 'data_input/array', 'data_input/HTML', 'data_input/JSON', 'data_input/XML'],function (require) {
     'use strict';
     
     var $, jquery;
@@ -13,7 +13,7 @@ define(['require', 'jquery', 'data_input/liveHTML', 'data_input/array', 'data_in
         "JSON": require('data_input/JSON'),
         "array": require('data_input/array'),
         "HTML": require('data_input/HTML'),
-        "xml": require('data_input/xml')
+        "XML": require('data_input/XML')
     };
     
     /**
