+++ tests/test.input.XML.html
@@ -26,13 +26,12 @@
                           <name>Item 1</name>
                           <searchable>Item 1</searchable>
                           <selected>false</selected>
-                          <icon>fa faIconClasses</icon>
                           <item>
                                   <name>Item 1.1</name>
                                   <value>Item 1.1</value>
                                   <searchable>Item 1.1</searchable>
                                   <selected>true</selected>
-                                  <image>Path/ToImage3</image>
+                                  <image>./quagga.jpg</image>
                           </item>
                           <item>
                                   <name>Item 1.2</name>
